Change color scale for magnetic map.
Shillington 2015 seems to have a scale from -200 to 200 nT. But the
data seems to have much more range than that. I'm not really sure what
the right thing to do here is. -500 to 500 seems to make more stuff
pop so I'll go with that. Also removed the grey fill for the rupture
zones.