Choose model profile start and end positions.
Project the nodes to the line defined by the first and last airgun
shot in order to determine the max offset, then use that information
to create the start of the model line (the end is the farthest airgun
shot). Also add a quick map of the model area and profile. Needed to
add geographiclib.